Remove unnecessary and dangerous dependency
You are not using model-build-features, which is NOT GOOD
for next Tizen versions as we are going to make it configurable
and accerlate build procedures. Device-specific adaptation
drivers and OAL instances are safe to use model-build-features.

If private forks are using model-build-features, you may
want to use private-specific macros there.

This commit is not required for Tizen 3.0. (It's for > 3.0)
If you do not feel comfortable for Tizen 3.0 release schedule,
you may delay merging, or delay SR, or put this into sandbox
branch.

Change-Id: I6431e09d017a39d7f1707b5ca88bf173c420143e
Signed-off-by: MyungJoo Ham <myungjoo.ham@samsung.com>